front: remove driver train schedule
