+++ front/src/applications/operationalStudies/types.ts
@@ -148,3 +148,17 @@ export type SimulationResultsData = {
   pathLength?: number;
   path?: PathfindingResultSuccess;
 };
+
+export type OperationalPointWithTimeAndSpeed = {
+  id: string | null;
+  name: string | null;
+  position: number;
+  speed: number;
+  time: number;
+  duration: number;
+  line_code: number | null;
+  track_number: number | null;
+  line_name: string | null;
+  track_name: string | null;
+  ch?: string | null;
+};
