+++ front/src/modules/simulationResult/SimulationResultExport/__tests__/utils.spec.ts
