+++ front/src/modules/simulationResult/SimulationResultExport/utils.ts
@@ -3,6 +3,7 @@ import { uniq } from 'lodash';
 
 import type { TrackSectionEntity } from 'applications/editor/tools/trackEdition/types';
 import type {
+  OperationalPointWithTimeAndSpeed,
   PathPropertiesFormatted,
   SimulationResponseSuccess,
 } from 'applications/operationalStudies/types';
@@ -13,70 +14,17 @@ import {
   type TrackSection,
   type TrainScheduleBase,
 } from 'common/api/osrdEditoastApi';
-import type { PositionSpeedTime, SpeedRanges } from 'reducers/simulationResults/types';
+import type { SpeedRanges } from 'reducers/simulationResults/types';
 import { store } from 'store';
 import { mmToM, msToKmhRounded } from 'utils/physics';
 import { ISO8601Duration2sec, ms2sec } from 'utils/timeManipulation';
-
-import type { OperationalPointWithTimeAndSpeed } from './types';
 
 export function massWithOneDecimal(number: number) {
   return Math.round(number / 100) / 10;
 }
 
-export function getTime(sec: number) {
-  const timeplus = new Date(sec * 1000);
-  const time = timeplus.toISOString().substr(11, 8);
-  const sixthDigit = parseInt(time[6], 10);
-  const seventhDigit = parseInt(time[7], 10);
-  if (sixthDigit >= 0 && sixthDigit < 2) {
-    if (sixthDigit === 1) {
-      if (seventhDigit <= 4) {
-        return time.slice(0, 5);
-      }
-      return `${time.slice(0, 5)}+`;
-    }
-    return time.slice(0, 5);
-  }
-  if (sixthDigit >= 1 && sixthDigit < 5) {
-    if (sixthDigit === 4) {
-      if (seventhDigit <= 4) {
-        return `${time.slice(0, 5)}+`;
-      }
-      if (seventhDigit > 4) {
-        timeplus.setMinutes(timeplus.getMinutes() + 1);
-        timeplus.setSeconds(0);
-        return timeplus.toISOString().substr(11, 8).slice(0, 5);
-      }
-    }
-    return `${time.slice(0, 5)}+`;
-  }
-  timeplus.setMinutes(timeplus.getMinutes() + 1);
-  timeplus.setSeconds(0);
-  return timeplus.toISOString().substr(11, 8).slice(0, 5);
-}
-
-// On the next 3 functions, we need to check if the found index is included in the array
+// On the next function, we need to check if the found index is included in the array
 // to prevent a white screen when datas are computing and synchronizing when switching the selected train
-export function getActualSpeedLeft(givenPosition: number, speed: PositionSpeedTime[]) {
-  const speedPosition = d3.bisectLeft(
-    speed.map((d) => d.position),
-    givenPosition
-  );
-  if (speedPosition >= speed.length - 1) {
-    return speed[speed.length - 1].speed * 3.6;
-  }
-  return speed[speedPosition].speed * 3.6;
-}
-
-export function getActualPositionLeft(givenPosition: number, speed: PositionSpeedTime[]) {
-  const speedPosition = d3.bisectLeft(
-    speed.map((d) => d.position),
-    givenPosition
-  );
-  if (speedPosition >= speed.length - 1) return speed[speed.length - 1].position / 1000;
-  return speed[speedPosition].position / 1000;
-}
 
 /**
  * Get the Vmax at a givenPosition (in meters), using vmax (MRSP in m/s)
@@ -105,90 +53,6 @@ export function getActualVmax(givenPosition: number, vmax: SpeedRanges) {
   const actualVMax = findActualVmax(givenPosition, vmax);
   return msToKmhRounded(actualVMax);
 }
-
-export function getActualSpeedRight(givenPosition: number, speed: PositionSpeedTime[]) {
-  const speedPosition =
-    d3.bisectLeft(
-      speed.map((d) => d.position),
-      givenPosition
-    ) - 1;
-  return speedPosition <= 0 ? speed[0].speed * 3.6 : speed[speedPosition].speed * 3.6;
-}
-
-export function getActualPositionRight(givenPosition: number, speed: PositionSpeedTime[]) {
-  const speedPosition =
-    d3.bisectLeft(
-      speed.map((d) => d.position),
-      givenPosition
-    ) - 1;
-  return speedPosition <= 0 ? speed[0].position / 1000 : speed[speedPosition].position / 1000;
-}
-
-export function getActualSpeed(givenPosition: number, speed: PositionSpeedTime[]) {
-  const speedA = getActualSpeedRight(givenPosition, speed);
-  const speedB = getActualSpeedLeft(givenPosition, speed);
-  const posA = getActualPositionRight(givenPosition, speed);
-  const posB = getActualPositionLeft(givenPosition, speed);
-  if (speedA === 0 || speedB === 0) return 0;
-  const a = (speedB - speedA) / (posB - posA);
-  const b = speedA - a * posA;
-  return Math.round(a * (givenPosition / 1000) + b);
-}
-
-interface SpeedDistance {
-  distance: number;
-  speed: number;
-}
-
-export function getAverageSpeed(posA: number, posB: number, speedList: PositionSpeedTime[]) {
-  let totalDistance = 0;
-  const speedsAndDistances: SpeedDistance[] = [];
-  let averageSpeed = 0;
-
-  // Filter concerned speed by posA & posB (all speed sections between the two positions)
-  const concernedSpeeds = speedList.filter((item) => item.position >= posA && item.position < posB);
-
-  // When concernedSpeeds is empty or < 2, take nearest plateau speed
-  if (concernedSpeeds.length === 1) return Math.round(concernedSpeeds[0].speed * 3.6 * 10) / 10;
-  if (concernedSpeeds.length === 0) {
-    let lastKnownSpeedPosition = d3.bisectLeft(
-      speedList.map((step) => step.position),
-      posB
-    );
-    if (lastKnownSpeedPosition >= speedList.length) {
-      lastKnownSpeedPosition = speedList.length - 1;
-    }
-    return Math.round(speedList[lastKnownSpeedPosition].speed * 3.6 * 10) / 10;
-  }
-
-  // Get an array with speed along distance and set a sum of distance
-  concernedSpeeds.forEach((actualPosition, idx) => {
-    if (idx !== 0) {
-      const actualDistance = actualPosition.position - speedList[idx - 1].position;
-      speedsAndDistances.push({
-        speed: actualPosition.speed,
-        distance: actualDistance,
-      } as SpeedDistance);
-      totalDistance += actualDistance;
-    }
-  });
-
-  // Weight speed with distance ratio and sum it for average speed
-  speedsAndDistances.forEach((step) => {
-    averageSpeed += step.speed * (step.distance / totalDistance);
-  });
-
-  // Get average speed with decimal in km/h
-  return Math.round(averageSpeed * 3.6 * 10) / 10;
-}
-
-export const isEco = (train: TrainScheduleBase) => {
-  const { schedule, margins } = train;
-  return (
-    (schedule && schedule.some((item) => item.arrival !== null)) ||
-    (margins && margins.values.some((value) => value !== '0%' && value !== '0min/100km'))
-  );
-};
 
 /**
  * Interpolate a speed or time value at a given position when the operational point's position
@@ -245,10 +109,7 @@ export const formatOperationalPoints = async (
   simulatedTrain: SimulationResponseSuccess,
   train: TrainScheduleBase,
   infraId: number
-): Promise<{
-  base: OperationalPointWithTimeAndSpeed[];
-  finalOutput: OperationalPointWithTimeAndSpeed[];
-}> => {
+): Promise<OperationalPointWithTimeAndSpeed[]> => {
   // Get operational points metadata
   const trackIds = uniq(operationalPoints.map((op) => op.part.track));
   const trackSections = await store
@@ -262,16 +123,11 @@ export const formatOperationalPoints = async (
     .unwrap();
 
   // Format operational points
-  const formattedStops: {
-    base: OperationalPointWithTimeAndSpeed[];
-    finalOutput: OperationalPointWithTimeAndSpeed[];
-  } = { base: [], finalOutput: [] };
+  const formattedStops: OperationalPointWithTimeAndSpeed[] = [];
 
-  const { base, final_output } = simulatedTrain;
+  const { final_output } = simulatedTrain;
 
   operationalPoints.forEach((op) => {
-    const { time: baseTime, speed: baseSpeed } = getTimeAndSpeed(base, op);
-
     const { time: finalOutputTime, speed: finalOutputSpeed } = getTimeAndSpeed(final_output, op);
 
     // Get duration
@@ -313,14 +169,7 @@ export const formatOperationalPoints = async (
       ch: op.extensions?.sncf?.ch || null,
     };
 
-    formattedStops.base.push({
-      // time refers to the time elapsed since departure so we need to add it to the start time
-      time: convertDepartureTimeIntoSec(train.start_time) + ms2sec(baseTime),
-      speed: baseSpeed,
-      ...opCommonProp,
-    });
-
-    formattedStops.finalOutput.push({
+    formattedStops.push({
       time: convertDepartureTimeIntoSec(train.start_time) + ms2sec(finalOutputTime),
       speed: finalOutputSpeed,
       ...opCommonProp,
