+++ front/tests/pages/op-timetable-page-model.ts
@@ -25,8 +25,6 @@ class OperationalStudiesTimetablePage {
 
   readonly simulationMap: Locator;
 
-  readonly simulationDriverTrainSchedule: Locator;
-
   readonly timetableFilterButton: Locator;
 
   readonly timetableFilterButtonClose: Locator;
@@ -54,7 +52,6 @@ class OperationalStudiesTimetablePage {
     this.spaceTimeChart = page.locator('.space-time-chart-container');
     this.timeStopsDataSheet = page.locator('.time-stops-datasheet');
     this.simulationMap = page.locator('.simulation-map');
-    this.simulationDriverTrainSchedule = page.locator('.simulation-driver-train-schedule');
     this.timetableFilterButton = page.getByTestId('timetable-filter-button');
     this.timetableFilterButtonClose = page.getByTestId('timetable-filter-button-close');
     this.editTrainButton = page.getByTestId('edit-train');
@@ -103,9 +100,8 @@ class OperationalStudiesTimetablePage {
       this.manchetteSpaceTimeChart,
       this.speedSpaceChart,
       this.spaceTimeChart,
-      this.timeStopsDataSheet,
       this.simulationMap,
-      this.simulationDriverTrainSchedule,
+      this.timeStopsDataSheet,
     ];
     await Promise.all(
       simulationResultsLocators.map((simulationResultsLocator) =>
